requirements.txt
dearpygui>=1.3.1
numpy>=1.21.5
pickle5>=0.0.12
pygame>=2.1.2
PyOpenGL>=3.1.5
